--- src/app/pipes/pretty-date.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { PrettyDatePipe } from './pretty-date.pipe';
-
-describe('PrettyDatePipe', () => {
-  it('create an instance', () => {
-    const pipe = new PrettyDatePipe();
-    expect(pipe).toBeTruthy();
-  });
-});
