+++ src/app/shared/pipes/index.ts
@@ -1,0 +1,3 @@
+export { PrettyDatePipe } from './pretty-date.pipe';
+export { SpecialConditionTitlePipe } from './special-condition-title.pipe';
+export { ZoneTitlePipe } from './zone-title.pipe';
